Ex6/requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
numpy==1.21.0
opencv-python==4.5.3.56
paho-mqtt==1.5.1
pkg-resources==0.0.0
requests==2.25.1
urllib3==1.26.6
